added nuenchow lib

+++ src/Test.java
@@ -1,13 +1,15 @@
 import com.peter.crypto.numberfield2d.NumberField;
 import com.peter.crypto.numberfield2d.NumberFieldFactory;
 
+import java.math.BigInteger;
+
 
 public class Test
 {
     public static void main (String[] args) throws Exception
     {
         NumberField n =
-                NumberFieldFactory.countedSpiralRightDown (Byte.class, 5, "motherfucker");
+                NumberFieldFactory.countedSpiralRightDown (Double.class, 5, "motherfucker");
         System.out.println(n);
 
 //        BigInteger bi = PiDigitsHex.packedPi(0,1000);
